apps/Stk: Add support to handle modem restart and ICC Refresh
Uninstall the STK app, stop StkAppService and remove the idle text
when StkAppService receives the radio unavailable or radio off event.
This change also fixes the issue with not starting of StkMenu and
StkInput Activity.

Also, IdleModeText and SETUP_MENU need to be cleared when ICC status changes to
REFRESH_RESET or RESET_INIT. Hence sending the Icc Status change to STkApp

Change-Id: I597ea8ee2755183e2e6d402bd001bfb5eb99b31d
CRs-Fixed:270755